requirements.txt
pandas=="2.2.1"
plotly-express=="0.4.1"
streamlit=="1.32.2"
psutil=="5.9.8"
